refactor: Remove unnecessary things
