+++ .env.example
@@ -26,12 +26,9 @@ RETRY_COUNT=3                              # Number of retry attempts on failure
 # ========================
 # AWS Configuration
 # ========================
-AWS_REGION=us-east-1                         # AWS region (e.g., us-east-1, eu-west-1)
 AWS_S3_BUCKET=your-s3-bucket-name            # Name of your S3 bucket for storage
 AWS_DYNAMO_TABLE_BATCH_RECORDS=BatchRecords  # Name of your DynamoDB table
 AWS_DYNAMO_TABLE_RECORDS_STATUS=ReportStatus # Name of your DynamoDB table
-AWS_ACCESS_KEY_ID=your-access-key-id         # AWS Access Key ID
-AWS_SECRET_ACCESS_KEY=your-secret-access-key# AWS Secret Access Key
 AWS_BUCKET_NAME_CSV=digifarm-files
 AWS_BUCKET_NAME_NFT=digifarm-nfts
 AWS_SQS_QUEUE_URL=http://your-sqs-url/queue/queue-name
